requirements.txt
Pillow
numpy
opencv-python
pytesseract
python-dotenv
requests
